Share texture session lifecycle
