+++ docs/render-session-refactor.md
@@ -273,7 +273,7 @@ The lowest-risk order is:
 
 - [x] Generalize map attachment from one `texture_session` to one render target
       session.
-- [ ] Move common session lifecycle into shared render-session code.
+- [x] Move common session lifecycle into shared render-session code.
 - [ ] Add MapLibre-owned/offscreen texture sessions for readback-first server
       use.
 - [ ] Add `mln_texture_read_premultiplied_rgba8` on rendered texture sessions.
